README updated to document how to run test suite. Updated test suite to test FMTO core class methods - Core.initializeAdapter() and Core.initializeDevice()

+++ test/global.teardown.mjs
@@ -21,7 +21,10 @@ export default async function ()
     await global.__BROWSER_GLOBAL__.close()
 
     // Clean up the wsEndpoint file
-    await fs.rm(DIR, { recursive: true, force: true }, (error) => {
-        // NOOP
-    })
+    if (!process.env.FMTO_TEST_DONT_CLEAN)
+    {
+        await fs.rm(DIR, { recursive: true, force: true }, (error) => {
+            // NOOP
+        })
+    }
 }
